refactor: remove If component & datasetreport component unit test
